% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencyconverter.R
\name{currency_conversion}
\alias{currency_conversion}
\title{Currency conversion function.}
\usage{
currency_conversion(
  amount,
  from_currency,
  to_currency,
  percentage_change = 0,
  rounding_option = "none"
)
}
\arguments{
\item{amount}{The amount(s) to convert.}

\item{from_currency}{Currency/currencies to convert from.}

\item{to_currency}{Currency/currencies to convert to.}

\item{percentage_change}{Percentage change(s) in currency rate.}

\item{rounding_option}{Rounding option for converted amount.}
}
\value{
Converted amount(s).
}
\description{
Currency conversion function.
}
